requirements.txt
# Core
python-dotenv==1.0.0

# Web Scraping
playwright==1.48.0

# Base de datos (preparado para futuro)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Procesamiento de datos
pandas==2.1.3

# Exportación Excel (preparado para futuro)
openpyxl==3.1.2

# HTTP requests
requests==2.31.0

# Utilidades de fecha
python-dateutil==2.8.2